Replace old titlebar components with the new
