+++ packages/2019-disaster-game/src/components/atoms/TitleBar/index.js
@@ -1,120 +1,43 @@
-import { memo, useState, useEffect, Fragment } from "react";
-import { connect } from "react-redux";
 /** @jsx jsx */
 import { css, jsx } from "@emotion/core";
-import PropTypes from "prop-types";
 
-import { getActiveChapterId } from "../../../state/chapters";
-import { TASKS } from "../../../constants/chapters";
-import QRCode from "../../../../assets/earthquake-heroes-qr-code.svg";
+import QRCodeSVG from "../../../../assets/earthquake-heroes-qr-code.svg";
+import Palette from "../../../constants/style";
+import PlayerStats from "./PlayerStats";
 import JourneyBar from "./JourneyBar";
-import SavedBar from "./SavedBar";
 
-const ContainerStyle = css`
+const containerStyle = css`
   position: absolute;
   top: 0;
   left: 0;
-  width: 100%;
+  width: calc(100% - 70px);
   z-index: 103;
-  pointer-events: none;
-  transform: translateY(-100%);
-  transition: transform 1s;
+  background-color: ${Palette.darkGrey};
+  height: 230px;
+  padding: 0 35px;
   display: grid;
-  grid-template-rows: auto;
-  padding-top: 25px;
-`;
-
-const onScreenStyle = css`
-  transform: translateY(0%);
-`;
-
-const infoContainer = css`
-  /* 40px padding on each side */
-  width: calc(100% - 80px);
-  padding: 0 40px;
-  display: grid;
-  grid-template-columns: 355px 2360px 1025px;
-  align-items: center;
-  margin-top: 10px;
-`;
-
-const fullInfoContainer = css`
-  grid-template-rows: repeat(2, auto);
-  grid-template-columns: 355px 3380px;
+  grid-template-columns: repeat(3, auto);
 `;
 
 const QRCodeStyle = css`
   height: 160px;
+  margin-top: 35px;
+  position: relative;
+  display: inline-grid;
 `;
 
-const TitleBar = ({ activeChapterId }) => {
-  const [open, setOpen] = useState(false);
-  // eslint-disable-next-line no-unused-vars
-  const [badgeDrawerOpen, setBadgeDrawerOpen] = useState(false);
-  const [startAnimateTimeout, setStartAnimateTimeout] = useState(null);
-  const [endAnimateTimeout, setEndAnimateTimeout] = useState(null);
-
-  const openBadgeDrawer = () => {
-    const newStartTimeout = setTimeout(() => {
-      setBadgeDrawerOpen(true);
-    }, 4000);
-    const newEndTimeout = setTimeout(() => {
-      setBadgeDrawerOpen(false);
-    }, 10000);
-    setStartAnimateTimeout(newStartTimeout);
-    setEndAnimateTimeout(newEndTimeout);
-  };
-
-  useEffect(() => {
-    setOpen(true);
-
-    return () => {
-      setOpen(false);
-      setBadgeDrawerOpen(false);
-      if (startAnimateTimeout) clearTimeout(startAnimateTimeout);
-      if (endAnimateTimeout) clearTimeout(endAnimateTimeout);
-    };
-  }, [endAnimateTimeout, startAnimateTimeout]);
-
+const TitleBar = () => {
   return (
-    <div
-      css={css`
-        ${ContainerStyle}
-        ${open && onScreenStyle}
-      `}
-    >
-      <div
-        css={css`
-          ${infoContainer};
-          ${badgeDrawerOpen && fullInfoContainer}
-        `}
-      >
-        <img
-          src={QRCode}
-          alt="QR code for civicplatform.org/EarthquakeHeroes"
-          css={QRCodeStyle}
-        />
-        <JourneyBar
-          badgeDrawerOpen={badgeDrawerOpen}
-          openBadgeDrawer={openBadgeDrawer}
-        />
-        {activeChapterId === TASKS && (
-          <Fragment>
-            {badgeDrawerOpen && <div />}
-            <SavedBar justifyRight={badgeDrawerOpen} />
-          </Fragment>
-        )}
-      </div>
+    <div css={containerStyle}>
+      <img
+        src={QRCodeSVG}
+        alt="QR code for civicplatform.org/EarthquakeHeroes"
+        css={QRCodeStyle}
+      />
+      <JourneyBar />
+      <PlayerStats />
     </div>
   );
 };
 
-TitleBar.propTypes = {
-  activeChapterId: PropTypes.string
-};
-
-const mapStateToProps = state => ({
-  activeChapterId: getActiveChapterId(state)
-});
-
-export default connect(mapStateToProps)(memo(TitleBar));
+export default TitleBar;
